+++ code/ui/tabs/types/createTabs.d.ts
@@ -1,6 +1,75 @@
-export * from './createTabs';
-export * from './StyledContext';
-export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import("@tamagui/web").GetFinalProps<import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
+import type { GroupProps } from '@tamagui/group';
+import { type RovingFocusGroupProps } from '@tamagui/roving-focus';
+import type { GetProps, TamaguiElement } from '@tamagui/web';
+import * as React from 'react';
+import type { LayoutRectangle } from 'react-native';
+import { TabsContentFrame as DefaultTabsContentFrame, TabsFrame as DefaultTabsFrame, TabsTriggerFrame as DefaultTabsTriggerFrame } from './Tabs';
+export declare function createTabs({ ContentFrame, TriggerFrame, TabsFrame, }: {
+    ContentFrame?: import("@tamagui/web").TamaguiComponent<import("@tamagui/web").TamaDefer, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
+        elevation?: number | import("@tamagui/web").SizeTokens | undefined;
+        inset?: number | import("@tamagui/web").SizeTokens | {
+            top?: number;
+            bottom?: number;
+            left?: number;
+            right?: number;
+        } | null | undefined;
+        transparent?: boolean | undefined;
+        fullscreen?: boolean | undefined;
+        circular?: boolean | undefined;
+        hoverTheme?: boolean | undefined;
+        pressTheme?: boolean | undefined;
+        focusTheme?: boolean | undefined;
+        elevate?: boolean | undefined;
+        bordered?: number | boolean | undefined;
+        backgrounded?: boolean | undefined;
+        radiused?: boolean | undefined;
+        padded?: boolean | undefined;
+        chromeless?: boolean | "all" | undefined;
+    }, import("@tamagui/web").StaticConfigPublic> | undefined;
+    TriggerFrame?: import("@tamagui/web").TamaguiComponent<import("@tamagui/web").TamaDefer, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
+        disabled?: boolean | undefined;
+        elevation?: number | import("@tamagui/web").SizeTokens | undefined;
+        inset?: number | import("@tamagui/web").SizeTokens | {
+            top?: number;
+            bottom?: number;
+            left?: number;
+            right?: number;
+        } | null | undefined;
+        transparent?: boolean | undefined;
+        fullscreen?: boolean | undefined;
+        circular?: boolean | undefined;
+        hoverTheme?: boolean | undefined;
+        pressTheme?: boolean | undefined;
+        focusTheme?: boolean | undefined;
+        elevate?: boolean | undefined;
+        bordered?: number | boolean | undefined;
+        backgrounded?: boolean | undefined;
+        radiused?: boolean | undefined;
+        padded?: boolean | undefined;
+        chromeless?: boolean | "all" | undefined;
+        size?: import("@tamagui/web").SizeTokens | undefined;
+        active?: boolean | undefined;
+        unstyled?: boolean | undefined;
+    }, import("@tamagui/web").StaticConfigPublic> | undefined;
+    TabsFrame?: import("@tamagui/web").TamaguiComponent<import("@tamagui/web").TamaDefer, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
+        size?: import("@tamagui/web").SizeTokens | undefined;
+        elevation?: number | import("@tamagui/web").SizeTokens | undefined;
+        inset?: number | import("@tamagui/web").SizeTokens | {
+            top?: number;
+            bottom?: number;
+            left?: number;
+            right?: number;
+        } | null | undefined;
+        fullscreen?: boolean | undefined;
+        circular?: boolean | undefined;
+        hoverTheme?: boolean | undefined;
+        pressTheme?: boolean | undefined;
+        focusTheme?: boolean | undefined;
+        elevate?: boolean | undefined;
+        bordered?: number | boolean | undefined;
+        unstyled?: boolean | undefined;
+    }, import("@tamagui/web").StaticConfigPublic> | undefined;
+}): React.ForwardRefExoticComponent<Omit<import("@tamagui/web").GetFinalProps<import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
     size?: import("@tamagui/web").SizeTokens | undefined;
     elevation?: number | import("@tamagui/web").SizeTokens | undefined;
     inset?: number | import("@tamagui/web").SizeTokens | {
@@ -17,21 +86,7 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
     elevate?: boolean | undefined;
     bordered?: number | boolean | undefined;
     unstyled?: boolean | undefined;
-}>, keyof {
-    value?: string;
-    defaultValue?: string | undefined;
-    onValueChange?: ((value: string) => void) | undefined;
-    orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-    dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-    activationMode?: "automatic" | "manual";
-}> & {
-    value?: string;
-    defaultValue?: string | undefined;
-    onValueChange?: ((value: string) => void) | undefined;
-    orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-    dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-    activationMode?: "automatic" | "manual";
-} & import("react").RefAttributes<import("@tamagui/web").TamaguiElement>> & import("@tamagui/web").StaticComponentObject<Omit<import("@tamagui/web").GetFinalProps<import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
+}>, keyof TabsExtraProps<string>> & TabsExtraProps<string> & React.RefAttributes<TamaguiElement>> & import("@tamagui/web").StaticComponentObject<Omit<import("@tamagui/web").GetFinalProps<import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
     size?: import("@tamagui/web").SizeTokens | undefined;
     elevation?: number | import("@tamagui/web").SizeTokens | undefined;
     inset?: number | import("@tamagui/web").SizeTokens | {
@@ -48,28 +103,7 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
     elevate?: boolean | undefined;
     bordered?: number | boolean | undefined;
     unstyled?: boolean | undefined;
-}>, keyof {
-    value?: string;
-    defaultValue?: string | undefined;
-    onValueChange?: ((value: string) => void) | undefined;
-    orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-    dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-    activationMode?: "automatic" | "manual";
-}> & {
-    value?: string;
-    defaultValue?: string | undefined;
-    onValueChange?: ((value: string) => void) | undefined;
-    orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-    dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-    activationMode?: "automatic" | "manual";
-}, import("@tamagui/web").TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & {
-    value?: string;
-    defaultValue?: string | undefined;
-    onValueChange?: ((value: string) => void) | undefined;
-    orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-    dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-    activationMode?: "automatic" | "manual";
-}, import("@tamagui/web").StackStyleBase, {
+}>, keyof TabsExtraProps<string>> & TabsExtraProps<string>, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & TabsExtraProps<string>, import("@tamagui/web").StackStyleBase, {
     size?: import("@tamagui/web").SizeTokens | undefined;
     elevation?: number | import("@tamagui/web").SizeTokens | undefined;
     inset?: number | import("@tamagui/web").SizeTokens | {
@@ -104,28 +138,7 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         elevate?: boolean | undefined;
         bordered?: number | boolean | undefined;
         unstyled?: boolean | undefined;
-    }>, keyof {
-        value?: string;
-        defaultValue?: string | undefined;
-        onValueChange?: ((value: string) => void) | undefined;
-        orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-        dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-        activationMode?: "automatic" | "manual";
-    }> & {
-        value?: string;
-        defaultValue?: string | undefined;
-        onValueChange?: ((value: string) => void) | undefined;
-        orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-        dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-        activationMode?: "automatic" | "manual";
-    }, import("@tamagui/web").TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & {
-        value?: string;
-        defaultValue?: string | undefined;
-        onValueChange?: ((value: string) => void) | undefined;
-        orientation?: import("@tamagui/roving-focus").RovingFocusGroupProps["orientation"];
-        dir?: import("@tamagui/roving-focus").RovingFocusGroupProps["dir"];
-        activationMode?: "automatic" | "manual";
-    }, import("@tamagui/web").StackStyleBase, {
+    }>, keyof TabsExtraProps<string>> & TabsExtraProps<string>, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & TabsExtraProps<string>, import("@tamagui/web").StackStyleBase, {
         size?: import("@tamagui/web").SizeTokens | undefined;
         elevation?: number | import("@tamagui/web").SizeTokens | undefined;
         inset?: number | import("@tamagui/web").SizeTokens | {
@@ -144,7 +157,7 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         unstyled?: boolean | undefined;
     }, import("@tamagui/web").StaticConfigPublic];
 } & {
-    List: import("react").ForwardRefExoticComponent<Omit<import("@tamagui/core").RNTamaguiViewNonStyleProps, "elevation" | keyof import("@tamagui/web").StackStyleBase | "transparent" | "fullscreen" | "circular" | "hoverTheme" | "pressTheme" | "focusTheme" | "elevate" | "bordered" | "backgrounded" | "radiused" | "padded" | "chromeless" | "size" | "unstyled"> & import("@tamagui/web").WithThemeValues<import("@tamagui/web").StackStyleBase> & {
+    List: React.ForwardRefExoticComponent<Omit<import("@tamagui/core").RNTamaguiViewNonStyleProps, "elevation" | keyof import("@tamagui/web").StackStyleBase | "transparent" | "fullscreen" | "circular" | "hoverTheme" | "pressTheme" | "focusTheme" | "elevate" | "bordered" | "backgrounded" | "radiused" | "padded" | "chromeless" | "size" | "unstyled"> & import("@tamagui/web").WithThemeValues<import("@tamagui/web").StackStyleBase> & {
         elevation?: number | import("@tamagui/web").SizeTokens | undefined;
         inset?: number | import("@tamagui/web").SizeTokens | {
             top?: number;
@@ -211,8 +224,15 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         size?: any;
         unstyled?: boolean | undefined;
     }>> & import("@tamagui/group").GroupExtraProps & {
+        /**
+         * Whether to loop over after reaching the end or start of the items
+         * @default true
+         */
         loop?: boolean;
-    } & import("react").RefAttributes<import("@tamagui/web").TamaguiElement>>;
+    } & React.RefAttributes<TamaguiElement>>;
+    /**
+     * @deprecated Use Tabs.Tab instead
+     */
     Trigger: import("@tamagui/web").TamaguiComponent<Omit<import("@tamagui/web").GetFinalProps<import("@tamagui/core").RNTamaguiViewNonStyleProps, import("@tamagui/web").StackStyleBase, {
         disabled?: boolean | undefined;
         elevation?: number | import("@tamagui/web").SizeTokens | undefined;
@@ -334,12 +354,15 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         active?: boolean | undefined;
         unstyled?: boolean | undefined;
     }>> & {
+        /** The value for the tabs state to be changed to after activation of the trigger */
         value: string;
-        onInteraction?: (type: "hover" | "focus" | "select", layout: import("./createTabs").TabLayout | null) => void;
+        /** Used for making custom indicators when trigger interacted with */
+        onInteraction?: (type: InteractionType, layout: TabLayout | null) => void;
+        /** Disables setting the active theme when tab is active */
         disableActiveTheme?: boolean;
     } & {
         __scopeTabs?: string;
-    }, import("@tamagui/web").TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & Omit<import("@tamagui/core").RNTamaguiViewNonStyleProps, "disabled" | "elevation" | keyof import("@tamagui/web").StackStyleBase | "transparent" | "fullscreen" | "circular" | "hoverTheme" | "pressTheme" | "focusTheme" | "elevate" | "bordered" | "backgrounded" | "radiused" | "padded" | "chromeless" | "size" | "active" | "unstyled"> & import("@tamagui/web").WithThemeValues<import("@tamagui/web").StackStyleBase> & {
+    }, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & Omit<import("@tamagui/core").RNTamaguiViewNonStyleProps, "disabled" | "elevation" | keyof import("@tamagui/web").StackStyleBase | "transparent" | "fullscreen" | "circular" | "hoverTheme" | "pressTheme" | "focusTheme" | "elevate" | "bordered" | "backgrounded" | "radiused" | "padded" | "chromeless" | "size" | "active" | "unstyled"> & import("@tamagui/web").WithThemeValues<import("@tamagui/web").StackStyleBase> & {
         disabled?: boolean | undefined;
         elevation?: number | import("@tamagui/web").SizeTokens | undefined;
         inset?: number | import("@tamagui/web").SizeTokens | {
@@ -412,8 +435,11 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         active?: boolean | undefined;
         unstyled?: boolean | undefined;
     }>> & {
+        /** The value for the tabs state to be changed to after activation of the trigger */
         value: string;
-        onInteraction?: (type: "hover" | "focus" | "select", layout: import("./createTabs").TabLayout | null) => void;
+        /** Used for making custom indicators when trigger interacted with */
+        onInteraction?: (type: InteractionType, layout: TabLayout | null) => void;
+        /** Disables setting the active theme when tab is active */
         disableActiveTheme?: boolean;
     } & {
         __scopeTabs?: string;
@@ -563,12 +589,15 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         active?: boolean | undefined;
         unstyled?: boolean | undefined;
     }>> & {
+        /** The value for the tabs state to be changed to after activation of the trigger */
         value: string;
-        onInteraction?: (type: "hover" | "focus" | "select", layout: import("./createTabs").TabLayout | null) => void;
+        /** Used for making custom indicators when trigger interacted with */
+        onInteraction?: (type: InteractionType, layout: TabLayout | null) => void;
+        /** Disables setting the active theme when tab is active */
         disableActiveTheme?: boolean;
     } & {
         __scopeTabs?: string;
-    }, import("@tamagui/web").TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & Omit<import("@tamagui/core").RNTamaguiViewNonStyleProps, "disabled" | "elevation" | keyof import("@tamagui/web").StackStyleBase | "transparent" | "fullscreen" | "circular" | "hoverTheme" | "pressTheme" | "focusTheme" | "elevate" | "bordered" | "backgrounded" | "radiused" | "padded" | "chromeless" | "size" | "active" | "unstyled"> & import("@tamagui/web").WithThemeValues<import("@tamagui/web").StackStyleBase> & {
+    }, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & Omit<import("@tamagui/core").RNTamaguiViewNonStyleProps, "disabled" | "elevation" | keyof import("@tamagui/web").StackStyleBase | "transparent" | "fullscreen" | "circular" | "hoverTheme" | "pressTheme" | "focusTheme" | "elevate" | "bordered" | "backgrounded" | "radiused" | "padded" | "chromeless" | "size" | "active" | "unstyled"> & import("@tamagui/web").WithThemeValues<import("@tamagui/web").StackStyleBase> & {
         disabled?: boolean | undefined;
         elevation?: number | import("@tamagui/web").SizeTokens | undefined;
         inset?: number | import("@tamagui/web").SizeTokens | {
@@ -641,8 +670,11 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         active?: boolean | undefined;
         unstyled?: boolean | undefined;
     }>> & {
+        /** The value for the tabs state to be changed to after activation of the trigger */
         value: string;
-        onInteraction?: (type: "hover" | "focus" | "select", layout: import("./createTabs").TabLayout | null) => void;
+        /** Used for making custom indicators when trigger interacted with */
+        onInteraction?: (type: InteractionType, layout: TabLayout | null) => void;
+        /** Disables setting the active theme when tab is active */
         disableActiveTheme?: boolean;
     } & {
         __scopeTabs?: string;
@@ -691,16 +723,7 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         radiused?: boolean | undefined;
         padded?: boolean | undefined;
         chromeless?: boolean | "all" | undefined;
-    }>, keyof {
-        value: string;
-        forceMount?: true;
-    }> & {
-        value: string;
-        forceMount?: true;
-    }, import("@tamagui/web").TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & {
-        value: string;
-        forceMount?: true;
-    }, import("@tamagui/web").StackStyleBase, {
+    }>, keyof TabsContentExtraProps> & TabsContentExtraProps, TamaguiElement, import("@tamagui/core").RNTamaguiViewNonStyleProps & TabsContentExtraProps, import("@tamagui/web").StackStyleBase, {
         elevation?: number | import("@tamagui/web").SizeTokens | undefined;
         inset?: number | import("@tamagui/web").SizeTokens | {
             top?: number;
@@ -722,4 +745,65 @@ export declare const Tabs: import("react").ForwardRefExoticComponent<Omit<import
         chromeless?: boolean | "all" | undefined;
     }, import("@tamagui/web").StaticConfigPublic>;
 };
-//# sourceMappingURL=index.d.ts.map
+type TabsFrameProps = GetProps<typeof DefaultTabsFrame>;
+type TabsExtraProps<Tab = string> = {
+    /** The value for the selected tab, if controlled */
+    value?: string;
+    /** The value of the tab to select by default, if uncontrolled */
+    defaultValue?: Tab;
+    /** A function called when a new tab is selected */
+    onValueChange?: (value: Tab) => void;
+    /**
+     * The orientation the tabs are layed out.
+     * Mainly so arrow navigation is done accordingly (left & right vs. up & down)
+     * @defaultValue horizontal
+     */
+    orientation?: RovingFocusGroupProps['orientation'];
+    /**
+     * The direction of navigation between toolbar items.
+     */
+    dir?: RovingFocusGroupProps['dir'];
+    /**
+     * Whether a tab is activated automatically or manually. Only supported in web.
+     * @defaultValue automatic
+     * */
+    activationMode?: 'automatic' | 'manual';
+};
+type TabsProps<Tab = string> = TabsFrameProps & TabsExtraProps<Tab>;
+type TabsListFrameProps = GroupProps;
+type TabsListProps = TabsListFrameProps & {
+    /**
+     * Whether to loop over after reaching the end or start of the items
+     * @default true
+     */
+    loop?: boolean;
+};
+type InteractionType = 'select' | 'focus' | 'hover';
+type TabLayout = LayoutRectangle;
+type TabsTriggerFrameProps = GetProps<typeof DefaultTabsTriggerFrame>;
+/**
+ * @deprecated use `TabTabsProps` instead
+ */
+type TabsTriggerProps = TabsTriggerFrameProps & {
+    /** The value for the tabs state to be changed to after activation of the trigger */
+    value: string;
+    /** Used for making custom indicators when trigger interacted with */
+    onInteraction?: (type: InteractionType, layout: TabLayout | null) => void;
+    /** Disables setting the active theme when tab is active */
+    disableActiveTheme?: boolean;
+};
+type TabsTabProps = TabsTriggerProps;
+type TabsTriggerLayout = LayoutRectangle;
+type TabsContentFrameProps = GetProps<typeof DefaultTabsContentFrame>;
+type TabsContentExtraProps = {
+    /** Will show the content when the value matches the state of Tabs root */
+    value: string;
+    /**
+     * Used to force mounting when more control is needed. Useful when
+     * controlling animation with Tamagui animations.
+     */
+    forceMount?: true;
+};
+type TabsContentProps = TabsContentFrameProps & TabsContentExtraProps;
+export type { TabsProps, TabsListProps, TabsTriggerProps, TabsTriggerLayout, TabsTabProps, TabsContentProps, TabLayout, };
+//# sourceMappingURL=createTabs.d.ts.map
